feat: add custom story exceptions

+++ app/db/db_user.py
@@ -2,6 +2,7 @@ from app.db.hash import Hash
 from sqlalchemy.orm.session import Session
 from app.schemas import UserBase
 from app.db.models import DbUser
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: UserBase):
@@ -22,13 +23,22 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    # Handle exception
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    # Handle exception
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
     user.update(
         {
             DbUser.username: request.username,
@@ -41,6 +51,10 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    # Handle exception
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
     db.delete(user)
     db.commit()
